delete useless IsNonUniqueNameError function, and delete db_test.go
